Fix versions installed for CKAN and extensions
Upgrade CKAN to the latest patch release (2.0.4). Patch releases are
fully backwards compatible and contain important security and stability
patches. This should have no effect on current functionality.

More info about CKAN releases can be found here:

http://docs.ckan.org/en/latest/maintaining/upgrading/index.html#ckan-releases

For the harvest and spatial extensions, the `stable` branch should work
from CKAN 2.0.x to 2.2.x (and hopefully future CKAN versions), so it's
better to target that one.

It is not necessary to create the `harvest` user any more, an internal
site user is used.